ajuste estrutura atomic desing
