+++ src/App.tsx
@@ -1,16 +1,20 @@
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Login from './pages/login/login.page';
-import GlobalStyle from './styles/globals';
+import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom';
+import Login from './components/pages/login/login.page';
+import GlobalStyle from './components/atoms/styles/globals';
 
 function App() {
+  const logado = false;
+
   return (
     <BrowserRouter>
       <GlobalStyle />
       <Switch>
-        <Route path="/login">
-          <Login />
-        </Route>
-        <Route path="/home">home</Route>
+        <Route path="/login" component={Login} />
+        {logado ? (
+          <Route component={() => <div>home</div>} path="/" />
+        ) : (
+          <Redirect to="/login" />
+        )}
       </Switch>
     </BrowserRouter>
   );
