+++ src/components/atoms/button/button.component.tsx
@@ -7,7 +7,7 @@ import {
   withStyles,
 } from '@material-ui/core/styles';
 import React from 'react';
-import { Colors } from '../colors/colors';
+import { Colors } from '../colors/colors.enum';
 
 const styles: (theme: Theme) => StyleRules<string> = (theme) =>
   createStyles({
