+++ src/components/atoms/colors/colors.enum.tsx
@@ -1,0 +1,5 @@
+export enum Colors {
+  Primary = '#FF7755',
+  Inputbackground = '#f3f3f3',
+  White = '#FFFFFF',
+}
