--- src/components/atoms/colors/colors.tsx
@@ -1,7 +1,0 @@
-const Colors = {
-  Primary: '#FF7755',
-  Inputbackground: '#f3f3f3',
-  White: '#FFFFFF',
-};
-
-export { Colors };
