+++ src/components/atoms/fonts/poppins/Poppins-Thin.ttf
