+++ src/components/atoms/styles/globals.tsx
