+++ src/components/pages/area-estudante/area-estudante.page.tsx
@@ -1,0 +1,5 @@
+const AreaEstudante = () => {
+  return 'AreaEstudante';
+};
+
+export default AreaEstudante;
